+ ADT::Heap::Clear + Per component Abs function for vectors + insertion sorted list test scenario # IsPointerValid now available in release # LinkedAVLTree now working

+++ src/OrBinaryTree.cpp
@@ -201,7 +201,8 @@ void OrE::ADT::BinaryTree::DeleteAll(BinaryTreeNodeP _pNode)
 // ******************************************************************************** //
 OrE::ADT::BinaryTree::~BinaryTree()
 {
-	DeleteAll(m_pRoot);
+	if( m_pRoot )
+		DeleteAll( m_pRoot );
 }
 
 // *************************************EOF**************************************** //
